BoardConfig: Add support for building GKI kernel for QGKI config
As with the new specifications, the GKI kernel must be built along
side of QGKI (non-debug) kernel. Hence, specify "kernel-gki" as the
additional kernel that needs to be built for QGKI, include
the list of ramdisk modules for the same configuration (in addition
to building just the GKI kernel) and also specify the size of the
boot-gki.img partition (which is same as that of boot.img).

Change-Id: I6cfc71ba08e8cc50bb5012357cb5d51f10c5c1ef